Horror Game
You awaken to find yourself in a dark room, with no memory of how you got there. You see a window at the opposite end of the room. As you approach, you feel fatigued. Something is not right about this place. You look out the window to see that you are in a secluded mansion in the forest. Your tasks are to navigate the mansion, avoid danger, and ultimately escape. Good luck.